refactor: move LocationSearchResult mapper to domain layer
