+++ app/src/main/java/com/axehai/weatherscope/domain/model/LocationSearchResultMapper.kt
@@ -1,0 +1,9 @@
+package com.axehai.weatherscope.domain.model
+
+fun LocationSearchResult.toActiveLocation() = ActiveLocation(
+    latitude = latitude,
+    longitude = longitude,
+    name = name,
+    country = country,
+    source = LocationSource.SEARCH
+)
